Added intro section page and intro to footer. Fixed text in wordpress banner
