+++ app/src/app/components/footer/footer.component.html
@@ -1,33 +1,17 @@
 <footer id="footer-section">
-  <mat-toolbar class="footer-toolbar">
-    <div class="footer-content">
-    
-       <div class="footer-section">
-        <h4 class="mat-h3">About</h4>
-        <mat-list>
-          <a class="footer-link" mat-list-item href="#">Home</a>
-          <a class="footer-link" mat-list-item href="#">About</a>
-          <a class="footer-link" mat-list-item href="#">Contact</a>
-        </mat-list>
-      </div>
-      <div class="footer-section">
-        <h4 class="mat-h3">Resources</h4>
-        <mat-list>
-          <a class="footer-link" mat-list-item href="#">Documentation</a>
-          <a class="footer-link" mat-list-item href="#">Data Structure</a>
-          <a class="footer-link" mat-list-item href="#">Support</a>
-        </mat-list>
-      </div>
-      <div class="footer-section">
-        <h4 class="mat-h3">Connect</h4>
-        <app-social-links></app-social-links>
-      </div>
-    </div>
-  </mat-toolbar>
-  <mat-toolbar color="primary" class="copyright-toolbar">
+  <div class="footer-content">
+    <span class="heading">My creative journey</span>
+     <p>
+      I am a full-stack developer and machine learning enthusiast with extensive experience across web development, cloud infrastructure, embedded systems, and research. My expertise enables me to contribute to both product development and academic projects.<br>
+
+With 5+ years of experience in .Net (MCV and Core), Angular, Django, and NestJS, I specialize in building scalable backend solutions. My work also includes 3+ years of A/B testing using Convert and VWO to enhance user experiences through data-driven optimization.<br>
+      </p>
+      <app-social-links></app-social-links>
+  </div>
+  <div color="primary" class="copyright-toolbar">
     <span>&copy; 2024 Zohair Mehtab. All rights reserved.</span>
     <!--<span class="toolbar-spacer"></span>
     <a class="footer-link" mat-button href="#">Privacy Policy</a>
     <a class="footer-link" mat-button href="#">Terms of Service</a>-->
-  </mat-toolbar>
+  </div>
 </footer>
