+++ app/src/app/components/footer/footer.component.scss
@@ -1,36 +1,29 @@
-footer {
+#footer-section {
+    padding: 32px 16px;
 
-    .footer-toolbar {
-        padding: 32px 16px;
-        height: auto;
-    }
-    
     .footer-content {
         display: flex;
-        flex-wrap: wrap;
-        justify-content: space-between;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
         width: 100%;
         max-width: 1200px;
         margin: 0 auto;
+        text-align: center;
     }
     
-    .footer-section {
-        flex: 1;
-        min-width: 200px;
-        margin-bottom: 24px;
+    .heading {
+        font-size: 1.2rem;
+        margin-bottom: 10px;
+        font-weight: bold;
     }
-    
-    .mat-h3 {
-        margin-bottom: 16px;
-    }
-
     
     .copyright-toolbar {
         font-size: 0.9em;
-        padding: 8px 16px;
         align-items: center;
         justify-content: center;
         text-align: center;
+        margin-top: 20px;
     }
     
     .toolbar-spacer {
