+++ app/src/app/components/header/header.component.html
@@ -17,9 +17,9 @@
                         <button class="nav-link" (click)="goToPage(item.name, item.id)">{{item.name}}</button>
                     </li>
                 }
-                <li class="nav-item">
+                <!--<li class="nav-item">
                     <a class="nav-link" href="#">Contact</a>
-                </li>
+                </li>-->
             </ul>
             <div class="authentication-buttons">
                 <button mat-raised-button id="sign-in-button" class="authentication-button">Sign In</button>
