+++ app/src/app/components/intro-banner/intro-banner.component.scss
@@ -27,6 +27,28 @@
     }
 }
 
+@for $i from 1 to 100 {
+    @keyframes progress-#{$i} {
+        0% {
+            width: $i - 1%;
+        }
+        100% {
+            width: $i * 1%;
+        }
+    }   
+
+    @-webkit-keyframes progress-#{$i} {
+        0% {
+            width: $i - 1%;
+        }
+        100% {
+            width: $i * 1%;
+        }
+    }
+}
+
+
+
 #intro-banner {
     width: 100%;
     height: 100vh;
@@ -53,19 +75,14 @@
             font-size: 24px;
             
         }
-
-        app-social-links {
-            justify-content: center;
-            align-items: center;
-            font-size: 25px;
-            display: flex;
-
-
-            .social-icons i {
-                margin-right: 20px;
-            }
-        }
         
+    }
+
+    app-social-links {
+        justify-content: center;
+        align-items: center;
+        font-size: 25px;
+        display: flex;
     }
 
     .intro-banner-content {
@@ -75,6 +92,7 @@
         max-width: 1000px;
         position: relative;
         width: 100%;
+        min-height: 300px;
         
     }
 
@@ -106,6 +124,13 @@
         }
     }
 
+    .loading-message {
+        width: 100%;
+        text-align: center;
+        font-size: 1.5rem;
+        margin: auto 20px 10px 20px;
+    }
+
     .loader {
         width: 100%;
         max-width: 1000px;
@@ -113,13 +138,21 @@
         height: 30px;
         border-radius: 10px;
         background-color: #fff;
-        margin: auto 20px 10% 20px;
+        margin: 10px 20px 10% 20px;
         position: relative;
 
         .progress {
             width: 1%;
             height: 30px;
             border-radius: 10px;
+
+            @for $i from 1 through 100 {
+                &[progress='#{$i}%'] {
+                    animation: progress-#{$i} 1s linear;
+                    -webkit-animation: progress-#{$i} 1s linear;
+                    width: $i * 1%;
+                }
+            }
         }
         .progress-text {
             position: absolute;
