+++ app/src/app/components/social-links/social-links.component.html
@@ -1,4 +1,7 @@
 <div class="social-icons">
+    <a mat-icon-button href="mailto:zuhairmhtb@gmail.com" aria-label="Medium">
+        <i class="fa-brands fa-google"></i>
+    </a>
     <a mat-icon-button href="https://www.linkedin.com/in/zuhair-mehtab-2550bb8b/" target="_blank" aria-label="LinkedIn">
         <i class="fa-brands fa-linkedin"></i>
     </a>
