+++ app/src/app/components/social-links/social-links.component.scss
@@ -1,6 +1,6 @@
 .social-icons {
     display: flex;
-    gap: 8px;
+    gap: 15px;
 
     @media all and (max-width: 767px) {
         justify-content: center;
