--- app/src/app/components/webgl/webgl.component.html
@@ -1,3 +1,0 @@
-<div style="display: flex; justify-content: center; align-items: center; width: 100%; height: 400px;" #webglCanvasContainer>
-    <canvas id="webgl-canvas" style="background-color: #cf8787; width: 100%; height: 100%" #webglCanvas></canvas>
-</div>
